get_metrics and metric class.
This commit adds get_metrics functions which will return a Metric object according to the model type. Since we can have different and more than one metrics for each model, it would be a good idea to encapsulate all metrics into a Metric class. Python dict is used to identify different metric and keys should be defined in the global_strings.py.